Add getEmployeeById functionality

+++ src/main/java/com/cardona/restapidemo/repository/EmployeeRepository.java
@@ -2,6 +2,6 @@ package com.cardona.restapidemo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.cardona.restapidemo.model.Employee;
 
-public interface EmployeeRepository extends JpaRepository<Employee,Long> {
+public interface EmployeeRepository extends JpaRepository<Employee,Integer> {
 
 }
